Video #37: Locking down deleteItem video (end of module #5)

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -3,10 +3,11 @@ const jwt = require("jsonwebtoken");
 const { randomBytes } = require("crypto");
 const { promisify } = require("util");
 const { transport, makeANiceEmail } = require("../mail");
+const { hasPermission } = require("../utils");
 
 const Mutations = {
   async createItem(parent, args, ctx, info) {
-    //TODO: Check if they are logged in
+    //Check if they are logged in
     if (!ctx.request.userId) {
       throw new Error("You must be logged in to do that!");
     }
@@ -53,10 +54,17 @@ const Mutations = {
       {
         where
       },
-      `{id title}`
+      `{id title user { id }}`
     ); // Usually we would pass info as the 2nd arg here but that will resolve the mutation which we don't want we want to preform some checks so instead we put raw graphql
     // 2. Check if they own that item, or have the permissions
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission =>
+      ["ADMIN", "ITEMDELETE"].includes(permission)
+    );
 
+    if (!ownsItem && hasPermissions) {
+      throw new Error("You don't have permission to do that");
+    }
     // 3. Delete it!
     return ctx.db.mutation.deleteItem({ where }, info);
   },
@@ -208,6 +216,34 @@ const Mutations = {
 
     // 8. return the new user
     return updatedUser;
+  },
+  async updatePermissions(parent, args, ctx, info) {
+    // 1. Check if they are logged in
+    if (!ctx.request.userId) {
+      throw new Error("You must be logged in to do that!");
+    }
+    // 2. Query the current user
+    const user = await ctx.db.query.user(
+      { where: { id: ctx.request.userId } },
+      info
+    );
+    // 3. Check if they have permissions to do this
+    hasPermission(user, ["ADMIN", "PERMISSIONUPDATE"]);
+    // 4. Update the permissions
+    return ctx.db.mutation.updateUser(
+      {
+        data: {
+          permissions: {
+            set: args.permissions
+          }
+        },
+        where: {
+          // We are passing the userId as we may be updating somebody elses permissions
+          id: args.userId
+        }
+      },
+      info
+    );
   }
 };
